+++ CHANGELOG.md
@@ -37,7 +37,7 @@
 - #939 Add unit test for Project kernel
 - #949 Implemented using threadpool for outgoing messages 
 - #961 Add list_tables() and describe_table() functions
-
+- #967 Add bc.get_free_memory() function
 
 ## Bug Fixes
 - #774 fixed build issues with latest cudf 0.15 including updating from_cudf